requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pyyaml>=6.0
pillow>=10.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
tqdm>=4.66.0
pandas>=2.0.0
numpy>=1.24.0
tensorboard>=2.14.0
wandb>=0.15.0
matplotlib>=3.7.0
